+++ scala.novaextension/Scripts/main.js
@@ -1,4 +1,6 @@
-var metalsLanguageClient = null;
+let metalsLanguageClient = null;
+
+const javaHomeKey = "java.home";
 
 function metalsExecuteCommand(command, args, onSuccess) {
   if (!metalsLanguageClient) return;
@@ -8,18 +10,18 @@ function metalsExecuteCommand(command, args, onSuccess) {
       "workspace/executeCommand",
       {
         command: command,
-        arguments: args
-      }
+        arguments: args,
+      },
     )
     .then(
-      response => {
+      (response) => {
         if (onSuccess) {
           onSuccess(response);
         }
       },
-      error => {
+      (error) => {
         console.log(`Failed to execute Metals command: ${error}`);
-      }
+      },
     );
 }
 
@@ -27,13 +29,13 @@ const importBuild = () => metalsExecuteCommand("build-import");
 const runDoctor = () => metalsExecuteCommand("doctor-run");
 
 function notify(key, title, message) {
-  let notification = new NotificationRequest(key);
+  const notification = new NotificationRequest(key);
   notification.title = nova.localize(title);
   notification.body = nova.localize(message);
   nova.notifications.add(notification);
 }
 
-function startMetals() {
+function startMetals(javaHome) {
   if (metalsLanguageClient) {
     deactivateMetals();
     nova.subscriptions.remove(metalsLanguageClient);
@@ -41,9 +43,19 @@ function startMetals() {
 
   const localPath = `${nova.extension.path}/metals`;
 
+  let env;
+  if (javaHome) {
+    env = {
+      "JAVA_HOME": javaHome,
+    };
+  } else {
+    env = {};
+  }
+
   // Create the client
   const serverOptions = {
     path: localPath,
+    env,
   };
   const clientOptions = {
     // The set of document syntaxes for which the server is valid
@@ -52,8 +64,8 @@ function startMetals() {
     initializationOptions: {
       isHttpEnabled: true,
       doctorVisibilityProvider: true,
-      doctorProvider: "json"
-    }
+      doctorProvider: "json",
+    },
   };
   metalsLanguageClient = new LanguageClient(
     "scala",
@@ -77,11 +89,11 @@ function setupDoctor() {
 
   metalsLanguageClient.onNotification(
     "metals/executeClientCommand",
-    request => {
+    (request) => {
       if (request.command === "metals-doctor-run") {
         runDoctor();
       }
-    }
+    },
   );
 }
 
@@ -96,7 +108,7 @@ function deactivateMetals() {
   metalsLanguageClient = null;
 }
 
-function installMetals(version, onSuccess, onFailure) {
+function installMetals(version, javaHome, onSuccess, onFailure) {
   const args = [
     "coursier",
     "bootstrap",
@@ -111,23 +123,33 @@ function installMetals(version, onSuccess, onFailure) {
     `org.scalameta:metals_2.13:${version}`,
     "-o",
     "metals",
-    "-f"
+    "-f",
   ];
+
+  let env;
+  if (javaHome) {
+    env = {
+      "JAVA_HOME": javaHome,
+    };
+  } else {
+    env = {};
+  }
 
   const process = new Process("/usr/bin/env", {
     args,
-    cwd: nova.extension.path
+    cwd: nova.extension.path,
+    env,
   });
 
-  process.onStdout(line => {
+  process.onStdout((line) => {
     console.log("Coursier: " + line);
   });
 
-  process.onStderr(line => {
+  process.onStderr((line) => {
     console.log("Coursier ERR: " + line);
   });
 
-  process.onDidExit(result => {
+  process.onDidExit((result) => {
     if (result === 0) onSuccess();
     else onFailure();
   });
@@ -135,72 +157,92 @@ function installMetals(version, onSuccess, onFailure) {
   process.start();
 }
 
-function runSetupSteps() {
-    startMetals();
-    setupDoctor();
+function runSetupSteps(javaHome) {
+  startMetals(javaHome);
+  setupDoctor();
 }
 
-function installMetalsIfNotExists(afterInstalled) {
-  const metalsExists = nova.fs.access(`${nova.extension.path}/metals`, nova.fs.F_OK);  
+function installMetalsIfNotExists(javaHome, afterInstalled) {
+  const metalsExists = nova.fs.access(
+    `${nova.extension.path}/metals`,
+    nova.fs.F_OK,
+  );
 
   if (!metalsExists) {
     installMetals(
       "latest.release",
+      javaHome,
       afterInstalled,
-      () => notify("metals-install-failed", "Metals: Install failed", "Coursier must be installed and on the system path.")
+      () =>
+        notify(
+          "metals-install-failed",
+          "Metals: Install failed",
+          "Coursier must be installed and on the system path.",
+        ),
     );
   } else {
     afterInstalled();
   }
 }
 
-function updateMetals() {
+function updateMetals(javaHome) {
   deactivateMetals();
 
   installMetals(
     "latest.release",
+    javaHome,
     () => {
       notify(
         "metals-updated",
         "Metals updated successfully",
       );
-      runSetupSteps();
+      runSetupSteps(javaHome);
     },
     () => {
       notify(
         "metals-update-failed",
         "Metals update failed",
-        "Failed to update metals. Ensure that Coursier is installed and on the system path."
+        "Failed to update metals. Ensure that Coursier is installed and on the system path.",
       );
-    }
+    },
   );
+}
+
+function restartMetals() {
+  const javaHome = nova.config.get(javaHomeKey);
+  startMetals(javaHome);
 }
 
 function registerCommands() {
   nova.commands.register(
     "metals.importBuild",
-    _ => importBuild()
+    (_) => importBuild(),
   );
 
   nova.commands.register(
     "metals.update",
-    _ => updateMetals()
+    (_) => updateMetals(nova.config.get(javaHomeKey)),
   );
 
   nova.commands.register(
     "metals.doctor",
-    _ => runDoctor()
+    (_) => runDoctor(),
+  );
+
+  nova.commands.register(
+    "metals.restart",
+    (_) => restartMetals(),
   );
 
   function registerMetalsCommand(
     novaCommand,
     metalsCommand,
     args,
-    onSuccess
+    onSuccess,
   ) {
     nova.commands.register(
       novaCommand,
-      _ => metalsExecuteCommand(metalsCommand, args, onSuccess)
+      (_) => metalsExecuteCommand(metalsCommand, args, onSuccess),
     );
   }
 
@@ -211,25 +253,40 @@ function registerCommands() {
   registerMetalsCommand("metals.cleanCompile", "compile-clean");
   registerMetalsCommand("metals.resetNotifications", "reset-notifications");
   registerMetalsCommand("metals.restartBuildServer", "build-restart");
-  registerMetalsCommand("metals.disconnectFromOldBuildServer", "build-disconnect");
+  registerMetalsCommand(
+    "metals.disconnectFromOldBuildServer",
+    "build-disconnect",
+  );
   registerMetalsCommand("metals.resetWorkspace", "reset-workspace");
   registerMetalsCommand(
     "metals.listBuildTargets",
     "list-build-targets",
     null,
-    response =>
+    (response) =>
       notify(
         "build-targets",
         "Metals build targets",
-        response.join("\n")
-      )
+        response.join("\n"),
+      ),
+  );
+}
+
+function registerConfig() {
+  nova.config.onDidChange(
+    javaHomeKey,
+    (value) => updateMetals(value),
   );
 }
 
 exports.activate = function () {
   registerCommands();
+  registerConfig();
 
-  installMetalsIfNotExists(runSetupSteps);
+  const javaHome = nova.config.get(javaHomeKey);
+  installMetalsIfNotExists(
+    javaHome,
+    () => runSetupSteps(javaHome),
+  );
 };
 
 exports.deactivate = function () {