+++ scala.novaextension/CHANGELOG.md
@@ -1,1 +1,6 @@
-## Version 0.0
+## Version 0.3.0
+
+- moves `java.home` setting to the global extension configuration
+- updates `tree-sitter-scala` to 0.23.2
+- adds `with` and `throw` keyword highlights
+- adds `Metals: Restart` command